Rename old policy.alphabet.Alphabet into AlphabetCommonClasses and start using the new Alphabet classes
